Complete rewrite and down size on code.
I have rewritten or removed any files and/or code, some of it is more efficiant, others cleaner. For the sake of archival all of the previous files are in the "Legacy Code" folder. This build is completely stable and will automatically host any and all folders and files in the current directory (not its sub-directories) on a server.js file.